refactor: piece 움직임에 대한 예외를 piece 패키지로 이동
